+++ tests/RefDocGen.UnitTests/CodeElements/Tools/MemberIdTests.cs
@@ -25,7 +25,7 @@ public class MemberIdTests
     [Fact]
     public void Of_ReturnsCorrectData_ForMethodWithASingleParameter()
     {
-        var param = MockParameter(typeof(string));
+        var param = MockParameterData(typeof(string));
         var method = MockMethodData("Execute", typeof(void), [param]);
 
         MemberId.Of(method).ShouldBe("Execute(System.String)");
@@ -34,7 +34,7 @@ public class MemberIdTests
     [Fact]
     public void Of_ReturnsCorrectData_ForMethodWithASingleRefParameter()
     {
-        var param = MockParameter(typeof(string), true);
+        var param = MockParameterData(typeof(string), isByRef: true);
         var method = MockMethodData("Execute", typeof(void), [param]);
 
         MemberId.Of(method).ShouldBe("Execute(System.String@)");
@@ -46,7 +46,7 @@ public class MemberIdTests
         var explicitInterfaceType = Substitute.For<ITypeNameData>();
         explicitInterfaceType.Id.Returns("MyApp.MyInterface");
 
-        var param = MockParameter(typeof(MemberInfo));
+        var param = MockParameterData(typeof(MemberInfo));
         var method = MockMethodData("Execute", typeof(void), [param], explicitInterfaceType);
 
         MemberId.Of(method).ShouldBe("MyApp#MyInterface#Execute(System.Reflection.MemberInfo)");
@@ -55,10 +55,10 @@ public class MemberIdTests
     [Fact]
     public void Of_ReturnsCorrectData_ForMethodWithMultipleParameters()
     {
-        var param1 = MockParameter(typeof(string));
-        var param2 = MockParameter(typeof(Type));
-        var param3 = MockParameter(typeof(int).MakeByRefType(), true);
-        var param4 = MockParameter(typeof(double).MakeArrayType());
+        var param1 = MockParameterData(typeof(string));
+        var param2 = MockParameterData(typeof(Type));
+        var param3 = MockParameterData(typeof(int).MakeByRefType(), true);
+        var param4 = MockParameterData(typeof(double).MakeArrayType());
         var method = MockMethodData("Execute", typeof(void), [param1, param2, param3, param4]);
 
         string expectedId = "Execute(System.String,System.Type,System.Int32@,System.Double[])";
@@ -72,7 +72,7 @@ public class MemberIdTests
         var returnType = Substitute.For<ITypeNameData>();
         returnType.Id.Returns("MyApp.MyClass");
 
-        var param = MockParameter(typeof(MemberInfo));
+        var param = MockParameterData(typeof(MemberInfo));
 
         var operatorData = Substitute.For<IOperatorData>();
 
@@ -89,7 +89,7 @@ public class MemberIdTests
     [Fact]
     public void Of_ReturnsCorrectData_ForIndexer()
     {
-        var param = MockParameter(typeof(int));
+        var param = MockParameterData(typeof(int));
 
         var indexerData = Substitute.For<IIndexerData>();
 
@@ -131,7 +131,7 @@ public class MemberIdTests
     /// <param name="type">Type of the parameter.</param>
     /// <param name="isByRef">Is the parameter passed by reference?</param>
     /// <returns>Mocked <see cref="IParameterData"/> instance.</returns>
-    private IParameterData MockParameter(Type type, bool isByRef = false)
+    private IParameterData MockParameterData(Type type, bool isByRef = false)
     {
         var param = Substitute.For<IParameterData>();
         var typeNameData = new TypeNameData(type);
